i got confused while reading this code, so rename it more sensibly
